Customers crud work

+++ database/DDL.sql
@@ -150,7 +150,7 @@ CREATE TABLE IF NOT EXISTS Customers (
   addressCity VARCHAR(45) NOT NULL,
   addressState VARCHAR(45) NOT NULL,
   addressZip INT NOT NULL,
-  cusTotalSales INT NOT NULL DEFAULT 0,
+  cusTotalSales DECIMAL(10, 2) NOT NULL DEFAULT 0,
   RewardsTiers_rewardsTierId INT,
   PRIMARY KEY (customerID),
   INDEX fk_Customers_RewardsTiers1_idx (RewardsTiers_rewardsTierId ASC) VISIBLE,
